# Revert "Implement LogSubscriber and Notifications"