--- a/lib/view_component/base.rb
+++ b/lib/view_component/base.rb
@@ -4,7 +4,6 @@ require "action_view"
 require "active_support/configurable"
 require "view_component/collection"
 require "view_component/compile_cache"
-require "view_component/log_subscriber"
 require "view_component/previewable"
 require "view_component/slotable"
 
@@ -75,14 +74,12 @@ module ViewComponent
       # Assign captured content passed to component as a block to @content
       @content = view_context.capture(self, &block) if block_given?
 
-      instrument("render_template") do
-        before_render
+      before_render
 
-        if render?
-          render_template_for(@variant)
-        else
-          ""
-        end
+      if render?
+        render_template_for(@variant)
+      else
+        ""
       end
     ensure
       @current_template = old_current_template
@@ -168,18 +165,6 @@ module ViewComponent
     # that inhibits encapsulation & reuse.
     def request
       @request ||= controller.request
-    end
-
-    def instrument(action, &block)
-      if self.class.respond_to?(:counter_argument_present?) && self.class.counter_argument_present?
-        return yield
-      else
-        instrument_payload = {
-          virtual_path: virtual_path, component_name: self.class.name, identifier: self.class.identifier
-        }
-
-        ActiveSupport::Notifications.instrument("#{action}.view_component", instrument_payload, &block)
-      end
     end
 
     attr_reader :content, :view_context