… lib/view_component/log_subscriber.rb
… test/view_component/log_subscriber_test.rb