Reverts github/view_component#435

@nicolas-brousse I'm reverting #435 for now as it appears to have introduced some performance issues, at least based on our benchmarks:

before:
```
Comparison:

            component::   182965.1 i/s
            partial::    17118.3 i/s - 10.69x  (± 0.00) slower
```

after:

```
Comparison:
          component::    18380.1 i/s
            partial::    17635.6 i/s - same-ish: difference falls within error
```

@BlakeWilliams will follow up with more detail on the original PR ❤️